% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hydro_year.R
\name{add_hydro_year}
\alias{add_hydro_year}
\title{Add a hydrological year column to a data.table (by reference)}
\usage{
add_hydro_year(ref_dat, start_month = 10)
}
\arguments{
\item{ref_dat}{a \code{\link[data.table]{data.table}} for which to add hydro_year column}

\item{start_month}{numeric, which should be the first month of the hydro_year}
}
\value{
Nothing, adds column by reference.
}
\description{
Adds a hydrological year starting at a specific month (default = 10). The hydrological year
is identified by the starting month, e.g. 2002-10 to 2002-12 and 2003-01 to 2003-09
will have the same hydro_year=2002.
}
\details{
Requires that input data has column named "date".
}
\examples{


dat <- data.table(date = lubridate::ymd(c("2002-08-15", "2002-09-30", "2002-10-01", "2002-12-31",
                                          "2003-01-01", "2003-08-15", "2003-09-30", "2003-10-01")))
dat
add_hydro_year(dat)
dat
add_hydro_year(dat, start_month = 9)
dat
}
