% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_month_fct.R
\name{add_month_fct}
\alias{add_month_fct}
\title{Add a month factor column to a data.table (by reference)}
\usage{
add_month_fct(ref_dat, first_month = 1)
}
\arguments{
\item{ref_dat}{a \code{\link[data.table]{data.table}} for which to add month_fct column}

\item{first_month}{numeric, month which should be the first level (default: 1, i.e. Jan)}
}
\value{
Nothing, adds column by reference.
}
\description{
Adds the month abbreviations (Jan, Feb, ...) as a factor (so to keep the annual order),
with the option to define the first month (as factor level), so also hydrological years
are possible (Oct, Nov, Dec, Jan, ..). Requires that input data has column named "month",
and adds a column named "month_fct".
}
\examples{

dat <- data.table(month = sample(1:12, 30, replace = T))

add_month_fct(dat)
str(dat)

add_month_fct(dat, 10)
summary(dat$month_fct)

}
