% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_season_fct.R
\name{add_season_fct}
\alias{add_season_fct}
\title{Add a season factor column to a data.table (by reference)}
\usage{
add_season_fct(ref_dat, first_season = "DJF", add_n_days = F)
}
\arguments{
\item{ref_dat}{a \code{\link[data.table]{data.table}} for which to add season column}

\item{first_season}{character season, which should be the first level (default: "JJA")}

\item{add_n_days}{logical; if T, adds the number of days in each season (of a non-leap year)}
}
\value{
Nothing, adds column by reference.
}
\description{
Adds the four seasons DJF, SON, MAM, JJA as a factor (so to keep the annual order),
with the option to define the first season (factor level). Requires that input data
has column named "month", and adds a column named "season".
}
\examples{

dat <- data.table(month = sample(1:12, 30, replace = T))

add_season_fct(dat)
str(dat)

add_season_fct(dat, first_season = "JJA", add_n_days = T)
str(dat)
}
