% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_plus.R
\name{\%+\%}
\alias{\%+\%}
\title{String plus (similar to python + operator for strings)}
\usage{
x \%+\% y
}
\arguments{
\item{x, y}{String to concatenate}
}
\value{
The concatenated string.
}
\description{
String plus (similar to python + operator for strings)
}
\examples{
"Hello" \%+\% " " \%+\% "world!"
}
