% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_parallel_ubuntu.R
\name{init_parallel_ubuntu}
\alias{init_parallel_ubuntu}
\title{Initialize the parallel backend for Ubuntu}
\usage{
init_parallel_ubuntu(ncores = NULL)
}
\arguments{
\item{ncores}{The number of cores to use. Default NULL uses all but 1 core.}
}
\value{
Nothing.
}
\description{
Initialize the parallel backend for Ubuntu
}
