% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_parallel_ubuntu_fork.R
\name{init_parallel_ubuntu_fork}
\alias{init_parallel_ubuntu_fork}
\title{Initialize the parallel backend for Ubuntu (using forking)}
\usage{
init_parallel_ubuntu_fork(ncores = NULL)
}
\arguments{
\item{ncores}{The number of cores to use. Default NULL uses all but 1 core.}
}
\value{
Nothing.
}
\description{
Initialize the parallel backend for Ubuntu (using forking)
}
