% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_parallel_win10.R
\name{init_parallel_win10}
\alias{init_parallel_win10}
\title{Initialize the parallel backend for Windows 10}
\usage{
init_parallel_win10(ncores = NULL)
}
\arguments{
\item{ncores}{The number of cores to use. Default NULL uses all but 1 core.}
}
\value{
Nothing.
}
\description{
Initialize the parallel backend for Windows 10
}
\examples{

}
