% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_fct.R
\name{month_fct}
\alias{month_fct}
\title{Create a month factor column}
\usage{
month_fct(month, first_month = 1)
}
\arguments{
\item{month}{a vector of months (numeric)}

\item{first_month}{numeric, month which should be the first level (default: 1, i.e. Jan)}
}
\value{
Factor of same length as input.
}
\description{
Returns the month abbreviations (Jan, Feb, ...) as a factor (so to keep the annual order),
with the option to define the first month (as factor level), so also hydrological years
are possible (Oct, Nov, Dec, Jan, ..). Requires a numeric input vector of months.
}
\examples{

month <- sample(1:12, 30, replace = TRUE)
month
month_fct(month)
month_fct(month, 10) # note the level order

}
