% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_cut_fct.R
\name{num_cut_fct}
\alias{num_cut_fct}
\title{Convert factor from \code{\link{cut}} to numeric}
\usage{
num_cut_fct(x, digits)
}
\arguments{
\item{x}{factor, output from \code{\link{cut}}}

\item{digits}{optional, use \code{\link{round}} on the result with specific digits}
}
\value{
numeric vector of same length as x
}
\description{
After using \code{\link{cut}}, this function will convert the resulting
interval to numeric values, which correspond to the middle of the interval.
}
\examples{
x <- runif(10, 0, 3000)
x_cut <- cut(x, breaks = 0:3 * 1000)
num_cut_fct(x_cut)

# compare the differences:
data.frame(x, x_cut, num_cut_fct(x_cut))

}
