% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_sym.R
\name{pval_sym}
\alias{pval_sym}
\title{Create the p-value symbols from standard R summary functions}
\usage{
pval_sym(p.values)
}
\arguments{
\item{p.values}{numeric p.values (between 0 and 1)}
}
\value{
Vector of symbols.
}
\description{
Creates a vector of the same length as the input with the standard codes
for significance in R, as for example used in summary.lm
}
\examples{

pval <- c(runif(10, 0, 0.2))
cbind(pval, pval_sym(pval))
}
