% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_names.R
\name{safe_names}
\alias{safe_names}
\title{Make names file safe}
\usage{
safe_names(x)
}
\arguments{
\item{x}{a vector of months (character)}
}
\value{
Vector of same length
}
\description{
Replaces all non-alphanumeric characters with _ and removes all accents.
}
\examples{

safe_names("x is.not-Fòàèin!")

}
