% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_fct.R
\name{season_fct}
\alias{season_fct}
\title{Returns a season factor}
\usage{
season_fct(month, first_season = "DJF")
}
\arguments{
\item{month}{numeric month}

\item{first_season}{character season, which should be the first level (default: "JJA")}
}
\value{
Factor vector of the same length as input.
}
\description{
Creates a factor with the four seasons DJF, SON, MAM, JJA (so to keep the annual order),
with the option to define the first season (factor level), from a vector of numeric/integer months.
}
\examples{

month <- sample(1:12, 30, replace = TRUE)
month
season_fct(month)
str(season_fct(month))

season_fct(month, first_season = "JJA")
str(season_fct(month, first_season = "JJA"))
}
