% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pctl.R
\name{calc_pctl}
\alias{calc_pctl}
\title{Calculate percentiles}
\usage{
calc_pctl(x, p, prefix = "p", ...)
}
\arguments{
\item{x}{numeric vector}

\item{p}{same as probs in \code{\link{quantile}}}

\item{prefix}{prefix to add before percentile value (default: 'p')}

\item{...}{other arguments passed to \code{\link{quantile}}}
}
\value{
list of percentiles with names "prefixX", where X is the
  percentile (probs*100).
}
\description{
Wrapper around \code{\link{quantile}} that returns a list and better names.
Thus, also easier usable within j of data.table's.
}
\examples{
x <- rnorm(1000)
calc_pctl(x)
calc_pctl(x, prefix = "pctl")

}
